Move StateManager to StateManagerImpl
